backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-dotenv
SQLAlchemy
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
celery
redis
requests
loguru
